Bottleneck: FP32 logit casting wastes memory bandwidth. Hypothesis: BF16 cross_entropy is safe because softcap=15 bounds logits to [-15,15], preventing overflow in log-sum-exp. Evidence: BF16 handles exp(15)~3.3M fine; softmax subtracts max first so all exponents <=0.

+++ train.py
@@ -583,12 +583,12 @@ class GPT(nn.Module):
         x = norm(x)
 
         softcap = 15
-        logits = self.lm_head(x).float()
+        logits = self.lm_head(x)
         logits = softcap * torch.tanh(logits / softcap)
 
         if targets is not None:
             loss = F.cross_entropy(
-                logits.float().view(-1, logits.size(-1)),
+                logits.view(-1, logits.size(-1)),
                 targets.view(-1),
                 ignore_index=-1,
                 reduction=reduction,
